fixed compiler errors and added definitions file. Remove expressTypes, generalTypes and mongooseTypes file in near future
