+++ server/src/models/User.ts
@@ -2,7 +2,7 @@
 import mongoose, { Schema } from 'mongoose';
 
 // Import types
-import { IUser } from 'mongooseTypes';
+import { IUser } from 'definitions';
 
 const AddressSchema: mongoose.Schema = new Schema({
     company: String,
