+++ server/src/routes/toolbox/Shipping.ts
@@ -4,8 +4,9 @@
 
 // Dependencies
 import express from 'express';
-const router = express.Router();
 import Route from '../Route';
+
+const router = express.Router();
 
 export default class Shipping extends Route {
     constructor(path: string, app: express.Application) {
