+++ server/src/routes/toolbox/Todos.ts
@@ -4,8 +4,9 @@
 
 // Dependencies
 import express from 'express';
-const router = express.Router();
 import Route from '../Route';
+
+const router = express.Router();
 
 export default class Todos extends Route {
     constructor(path: string, app: express.Application) {
