+++ server/src/types/expressTypes.ts
@@ -2,20 +2,24 @@
  * @overview This file is custom types that needs to be added to the application for all express types
  */
 
-import { Router, Request, Response } from 'express';
+// import { Router, Request, Response } from 'express';
 
-/*** Router ***/
+// /*** Router ***/
 
-// This interface extends Router specifically for commercebox
-export interface IRouterExtended extends Router {
-    // Allow router to have an index call
-    [index: string]: any;
-}
+// // This interface extends Router specifically for commercebox
+// export interface IRouterExtended extends Router {
+//     // Allow router to have an index call
+//     [index: string]: any;
+// }
 
-/*** Request && Response ***/
+// /*** Request && Response ***/
 
-// Extends Request object
-export interface IRequestExtended extends Request {}
+// // Extends Request object
+// export interface IRequestExtended extends Request {
+//     user: {
+//         _id: Object
+//     }
+// }
 
-// Extends Response object
-export interface IResponseExtended extends Response {}
+// // Extends Response object
+// export interface IResponseExtended extends Response {}
