+++ server/src/utils/woocommerce/wooTools.ts
@@ -1,3 +1,7 @@
+import {
+    WoocommerceTokens
+} from 'definitions';
+
 const woocommerce = require('woocommerce-api');
 
 /**
